Travis CI integration, update to newer CUnit and fix cast errors.

+++ CContainers/Tests/TestPDoubleLinkedList.c
@@ -1,8 +1,8 @@
 #include "TestPDoubleLinkedList.h"
 #include <stdlib.h>
 
-#include <cunit/CUnit.h>
-#include <cunit/CUnitCI.h>
+#include <CUnit/CUnit.h>
+#include <CUnit/CUnitCI.h>
 
 #include <CContainers/PDoubleLinkedList.h>
 #include <CContainers/Utils.h>
@@ -15,7 +15,7 @@ void PDoubleLinkedListSuite_Helper_CheckSize (uint size)
 {
     if (size != PDoubleLinkedList_Size (pDoubleLinkedListHandle))
     {
-        printf ("\n        Size (e/a): %d, %d.", size, PDoubleLinkedList_Size (pDoubleLinkedListHandle));
+        printf ("\n        Size (e/a): %d, %d.", size, (int) PDoubleLinkedList_Size (pDoubleLinkedListHandle));
         CU_FAIL_FATAL ("Expected and actual values are not equal!");
     }
 }
